requirements.txt
pypdf==4.3.1
fpdf2==2.7.9
numpy==1.26.4
sentence-transformers==5.2.0
pytest==8.2.0
